# docs/openapi.json
{
    "openapi": "3.1.0",
    "info": {
        "title": "TM Seasons API Docs",
        "version": "1.0.0",
        "description": "REST API for TM Seasons, covering Discord authentication, Trackmania club and map administration, season lifecycle management, leaderboard polling, event-based scoring, standings, and activity feeds."
    },
    "servers": [
        {
            "url": "https://tm-bot.test/api",
            "description": "Local"
        }
    ],
...
    "security": [
        {
            "sanctum": []
        }
    ],
    "paths": {
...
        "/admin/maps/{map}": {
            "get": {
                "operationId": "admin.maps.show",
                "description": "Returns the selected imported map and admin-visible metadata.",
                "summary": "Get a map for admins",
                "tags": [
                    "Trackmania - Admin"
                ],
                "parameters": [
                    {
                        "name": "map",
                        "in": "path",
                        "required": true,
                        "description": "The map ID",
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "`MapResource`",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/components/schemas/MapResource"
                                        }
                                    },
                                    "required": [
                                        "data"
                                    ]
                                }
                            }
                        }
                    },
                    "404": {
                        "$ref": "#/components/responses/ModelNotFoundException"
                    },
                    "401": {
                        "$ref": "#/components/responses/AuthenticationException"
                    }
                }
            },
            "patch": {
                "operationId": "admin.maps.update",
                "description": "Updates editable fields for an imported Trackmania map.",
                "summary": "Update map metadata",
                "tags": [
                    "Trackmania - Admin"
                ],
                "parameters": [
                    {
                        "name": "map",
                        "in": "path",
                        "required": true,
                        "description": "The map ID",
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/UpdateMapRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "`MapResource`",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/components/schemas/MapResource"
                                        }
                                    },
                                    "required": [
                                        "data"
                                    ]
                                }
                            }
                        }
                    },
                    "404": {
                        "$ref": "#/components/responses/ModelNotFoundException"
                    },
                    "401": {
                        "$ref": "#/components/responses/AuthenticationException"
                    },
                    "422": {
                        "$ref": "#/components/responses/ValidationException"
                    }
                }
            },
            "delete": {
                "operationId": "admin.maps.destroy",
                "description": "Removes an imported map from the local database.",
                "summary": "Delete a map",
                "tags": [
                    "Trackmania - Admin"
                ],
                "parameters": [
                    {
                        "name": "map",
                        "in": "path",
                        "required": true,
                        "description": "The map ID",
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No content"
                    },
                    "404": {
                        "$ref": "#/components/responses/ModelNotFoundException"
                    },
                    "401": {
                        "$ref": "#/components/responses/AuthenticationException"
                    }
                }
            }
        },
...
        "/admin/seasons/{season}": {
            "get": {
                "operationId": "admin.seasons.show",
                "description": "Returns a season with creator and ordered map assignments.",
                "summary": "Get a season for admins",
                "tags": [
                    "Seasons - Admin"
                ],
                "parameters": [
                    {
                        "name": "season",
                        "in": "path",
                        "required": true,
                        "description": "The season ID",
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "`SeasonResource`",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/components/schemas/SeasonResource"
                                        }
                                    },
                                    "required": [
                                        "data"
                                    ]
                                }
                            }
                        }
                    },
                    "404": {
                        "$ref": "#/components/responses/ModelNotFoundException"
                    },
                    "401": {
                        "$ref": "#/components/responses/AuthenticationException"
                    }
                }
            },
            "patch": {
                "operationId": "admin.seasons.update",
                "description": "Updates season settings and validates the requested lifecycle state.",
                "summary": "Update a season",
                "tags": [
                    "Seasons - Admin"
                ],
                "parameters": [
                    {
                        "name": "season",
                        "in": "path",
                        "required": true,
                        "description": "The season ID",
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/UpdateSeasonRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "`SeasonResource`",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/components/schemas/SeasonResource"
                                        }
                                    },
                                    "required": [
                                        "data"
                                    ]
                                }
                            }
                        }
                    },
                    "404": {
                        "$ref": "#/components/responses/ModelNotFoundException"
                    },
                    "401": {
                        "$ref": "#/components/responses/AuthenticationException"
                    },
                    "422": {
                        "$ref": "#/components/responses/ValidationException"
                    }
                }
            },
            "delete": {
                "operationId": "admin.seasons.destroy",
                "description": "Removes the selected season from the local database.",
                "summary": "Delete a season",
                "tags": [
                    "Seasons - Admin"
                ],
                "parameters": [
                    {
                        "name": "season",
                        "in": "path",
                        "required": true,
                        "description": "The season ID",
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No content"
                    },
                    "404": {
                        "$ref": "#/components/responses/ModelNotFoundException"
                    },
                    "401": {
                        "$ref": "#/components/responses/AuthenticationException"
                    }
                }
            }
        },
        "/admin/seasons/{season}/finalize": {
            "post": {
                "operationId": "admin.seasons.finalize",
                "description": "Calculates final placement rewards and transitions an ended season to finalized.",
                "summary": "Finalize a season",
                "tags": [
                    "Seasons - Admin"
                ],
                "parameters": [
                    {
                        "name": "season",
                        "in": "path",
                        "required": true,
                        "description": "The season ID",
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {}
                                        }
                                    },
                                    "required": [
                                        "data"
                                    ]
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "message": {
                                            "type": "string"
                                        }
                                    },
                                    "required": [
                                        "message"
                                    ]
                                }
                            }
                        }
                    },
                    "404": {
                        "$ref": "#/components/responses/ModelNotFoundException"
                    },
                    "401": {
                        "$ref": "#/components/responses/AuthenticationException"
                    }
                }
            }
        },
        "/admin/seasons/update-statuses": {
            "post": {
                "operationId": "admin.seasons.updateStatuses",
                "description": "Applies scheduled lifecycle transitions based on the current time.",
                "summary": "Update automatic season statuses",
                "tags": [
                    "Seasons - Admin"
                ],
                "responses": {
                    "200": {
                        "description": "",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "object",
                                            "properties": {
                                                "activated": {
                                                    "type": "string"
                                                },
                                                "ended": {
                                                    "type": "string"
                                                }
                                            },
                                            "required": [
                                                "activated",
                                                "ended"
                                            ]
                                        }
                                    },
                                    "required": [
                                        "data"
                                    ]
                                }
                            }
                        }
                    },
                    "401": {
                        "$ref": "#/components/responses/AuthenticationException"
                    }
                }
            }
        },
...
        "/admin/seasons/{season}/maps/{map}": {
            "patch": {
                "operationId": "admin.seasons.maps.update",
                "description": "Changes ordering or active state for a map attached to a season.",
                "summary": "Update a season map assignment",
                "tags": [
                    "Seasons - Admin"
                ],
                "parameters": [
                    {
                        "name": "season",
                        "in": "path",
                        "required": true,
                        "description": "The season ID",
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "map",
                        "in": "path",
                        "required": true,
                        "description": "The map ID",
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/UpdateSeasonMapRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "`SeasonResource`",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/components/schemas/SeasonResource"
                                        }
                                    },
                                    "required": [
                                        "data"
                                    ]
                                }
                            }
                        }
                    },
                    "404": {
                        "$ref": "#/components/responses/ModelNotFoundException"
                    },
                    "401": {
                        "$ref": "#/components/responses/AuthenticationException"
                    },
                    "422": {
                        "$ref": "#/components/responses/ValidationException"
                    }
                }
            },
            "delete": {
                "operationId": "admin.seasons.maps.destroy",
                "description": "Removes a map assignment from the selected season.",
                "summary": "Detach a map from a season",
                "tags": [
                    "Seasons - Admin"
                ],
                "parameters": [
                    {
                        "name": "season",
                        "in": "path",
                        "required": true,
                        "description": "The season ID",
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "map",
                        "in": "path",
                        "required": true,
                        "description": "The map ID",
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No content"
                    },
                    "404": {
                        "$ref": "#/components/responses/ModelNotFoundException"
                    },
                    "401": {
                        "$ref": "#/components/responses/AuthenticationException"
                    }
                }
            }
        },
...
        "/admin/seasons/{season}/records": {
            "get": {
                "operationId": "admin.seasons.records",
                "description": "Returns paginated player records for a season, optionally filtered by map or player.",
                "summary": "List season map records",
                "tags": [
                    "Seasons - Admin"
                ],
                "parameters": [
                    {
                        "name": "season",
                        "in": "path",
                        "required": true,
                        "description": "The season ID",
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "map_id",
                        "in": "query",
                        "schema": {
                            "type": "integer",
                            "default": 0
                        }
                    },
                    {
                        "name": "player_id",
                        "in": "query",
                        "schema": {
                            "type": "integer",
                            "default": 0
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Array of `SeasonMapPlayerRecordResource`",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/SeasonMapPlayerRecordResource"
                                            }
                                        }
                                    },
                                    "required": [
                                        "data"
                                    ]
                                }
                            }
                        }
                    },
                    "404": {
                        "$ref": "#/components/responses/ModelNotFoundException"
                    },
                    "401": {
                        "$ref": "#/components/responses/AuthenticationException"
                    }
                }
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "sanctum": {
                "type": "http",
                "description": "Use a Laravel Sanctum bearer token for protected endpoints.",
                "scheme": "bearer"
            }
        },
        "schemas": {
...
            "MapResource": {
                "type": "object",
                "properties": {
                    "id": {
                        "type": "integer"
                    },
                    "uid": {
                        "type": "string"
                    },
                    "nadeo_map_id": {
                        "type": [
                            "string",
                            "null"
                        ]
                    },
                    "name": {
                        "type": [
                            "string",
                            "null"
                        ]
                    },
                    "author_account_id": {
                        "type": [
                            "string",
                            "null"
                        ]
                    },
                    "author_name": {
                        "type": [
                            "string",
                            "null"
                        ]
                    },
                    "author_time": {
                        "type": [
                            "integer",
                            "null"
                        ]
                    },
                    "gold_time": {
                        "type": [
                            "integer",
                            "null"
                        ]
                    },
                    "silver_time": {
                        "type": [
                            "integer",
                            "null"
                        ]
                    },
                    "bronze_time": {
                        "type": [
                            "integer",
                            "null"
                        ]
                    },
                    "map_type": {
                        "type": [
                            "string",
                            "null"
                        ]
                    },
                    "map_style": {
                        "type": [
                            "string",
                            "null"
                        ]
                    },
                    "thumbnail_url": {
                        "type": [
                            "string",
                            "null"
                        ]
                    },
                    "collection_name": {
                        "type": [
                            "string",
                            "null"
                        ]
                    },
                    "uploaded_at": {
                        "type": "string"
                    },
                    "source_updated_at": {
                        "type": "string"
                    },
                    "created_at": {
                        "type": "string"
                    },
                    "updated_at": {
                        "type": "string"
                    },
                    "season_pivot": {
                        "type": "object",
                        "properties": {
                            "id": {
                                "type": "string"
                            },
                            "order_index": {
                                "type": "string"
                            },
                            "is_active": {
                                "type": "boolean"
                            }
                        },
                        "required": [
                            "id",
                            "order_index",
                            "is_active"
                        ]
                    }
                },
                "required": [
                    "id",
                    "uid",
                    "nadeo_map_id",
                    "name",
                    "author_account_id",
                    "author_name",
                    "author_time",
                    "gold_time",
                    "silver_time",
                    "bronze_time",
                    "map_type",
                    "map_style",
                    "thumbnail_url",
                    "collection_name",
                    "uploaded_at",
                    "source_updated_at",
                    "created_at",
                    "updated_at"
                ],
                "title": "MapResource"
            },
...
            "UpdateMapRequest": {
                "type": "object",
                "properties": {
                    "name": {
                        "type": [
                            "string",
                            "null"
                        ],
                        "maxLength": 255
                    },
                    "author_name": {
                        "type": [
                            "string",
                            "null"
                        ],
                        "maxLength": 255
                    },
                    "map_type": {
                        "type": [
                            "string",
                            "null"
                        ],
                        "maxLength": 255
                    },
                    "map_style": {
                        "type": [
                            "string",
                            "null"
                        ],
                        "maxLength": 255
                    },
                    "collection_name": {
                        "type": [
                            "string",
                            "null"
                        ],
                        "maxLength": 255
                    },
                    "thumbnail_url": {
                        "type": [
                            "string",
                            "null"
                        ],
                        "format": "uri",
                        "maxLength": 2048
                    },
                    "author_time": {
                        "type": [
                            "integer",
                            "null"
                        ],
                        "minimum": 0
                    },
                    "gold_time": {
                        "type": [
                            "integer",
                            "null"
                        ],
                        "minimum": 0
                    },
                    "silver_time": {
                        "type": [
                            "integer",
                            "null"
                        ],
                        "minimum": 0
                    },
                    "bronze_time": {
                        "type": [
                            "integer",
                            "null"
                        ],
                        "minimum": 0
                    }
                },
                "title": "UpdateMapRequest"
            },
            "UpdateSeasonMapRequest": {
                "type": "object",
                "properties": {
                    "order_index": {
                        "type": "integer",
                        "minimum": 0
                    },
                    "is_active": {
                        "type": "boolean"
                    }
                },
                "title": "UpdateSeasonMapRequest"
            },
...
        },
        "responses": {
            "AuthenticationException": {
                "description": "Unauthenticated",
                "content": {
                    "application/json": {
                        "schema": {
                            "type": "object",
                            "properties": {
                                "message": {
                                    "type": "string",
                                    "description": "Error overview."
                                }
                            },
                            "required": [
                                "message"
                            ]
                        }
                    }
                }
            },
            "ModelNotFoundException": {
                "description": "Not found",
                "content": {
                    "application/json": {
                        "schema": {
                            "type": "object",
                            "properties": {
                                "message": {
                                    "type": "string",
                                    "description": "Error overview."
                                }
                            },
                            "required": [
                                "message"
                            ]
                        }
                    }
                }
            },
...
        }
    }
}